+++ R/get_ssurgo_tables.R
@@ -87,10 +87,10 @@ get_ssurgo_tables <- function(lonlat, shift = -1, aoi, verbose = FALSE){
 
   if(verbose == FALSE){
     ## res <- suppressWarnings(soilDB::SDA_spatialQuery(spg, what = 'mukey'))
-    res <- suppressWarnings(soilDB::SDA_spatialQuery(spg, what = 'geom', geomIntersection = TRUE))
+    res <- suppressWarnings(soilDB::SDA_spatialQuery(spg, what = 'mupolygon', geomIntersection = TRUE))
   }else{
     ## res <- soilDB::SDA_spatialQuery(spg, what = 'mukey')  
-    res <- soilDB::SDA_spatialQuery(spg, what = 'geom', geomIntersection = TRUE)
+    res <- soilDB::SDA_spatialQuery(spg, what = 'mupolygon', geomIntersection = TRUE)
   }
   
   mu.is <- soilDB::format_SQL_in_statement(res$mukey)